# core: improve Seek and optimise seek time for transfer logs