--- a/pkg/core/dao/dao.go
+++ b/pkg/core/dao/dao.go
@@ -42,7 +42,7 @@ type DAO interface {
 	GetCurrentHeaderHeight() (i uint32, h util.Uint256, err error)
 	GetHeaderHashes() ([]util.Uint256, error)
 	GetTokenTransferInfo(acc util.Uint160) (*state.TokenTransferInfo, error)
-	GetTokenTransferLog(acc util.Uint160, index uint32, isNEP11 bool) (*state.TokenTransferLog, error)
+	GetTokenTransferLog(acc util.Uint160, start uint64, index uint32, isNEP11 bool) (*state.TokenTransferLog, error)
 	GetStateSyncPoint() (uint32, error)
 	GetStateSyncCurrentBlockHeight() (uint32, error)
 	GetStorageItem(id int32, key []byte) state.StorageItem
@@ -57,7 +57,7 @@ type DAO interface {
 	PutContractID(id int32, hash util.Uint160) error
 	PutCurrentHeader(hashAndIndex []byte) error
 	PutTokenTransferInfo(acc util.Uint160, bs *state.TokenTransferInfo) error
-	PutTokenTransferLog(acc util.Uint160, index uint32, isNEP11 bool, lg *state.TokenTransferLog) error
+	PutTokenTransferLog(acc util.Uint160, start uint64, index uint32, isNEP11 bool, lg *state.TokenTransferLog) error
 	PutStateSyncPoint(p uint32) error
 	PutStateSyncCurrentBlockHeight(h uint32) error
 	PutStorageItem(id int32, key []byte, si state.StorageItem) error
@@ -180,21 +180,66 @@ func (dao *Simple) putTokenTransferInfo(acc util.Uint160, bs *state.TokenTransfe
 
 // -- start transfer log.
 
-func getTokenTransferLogKey(acc util.Uint160, index uint32, isNEP11 bool) []byte {
-	key := make([]byte, 1+util.Uint160Size+4)
+func getTokenTransferLogKey(acc util.Uint160, newestTimestamp uint64, index uint32, isNEP11 bool) []byte {
+	key := make([]byte, 1+util.Uint160Size+8+4)
 	if isNEP11 {
 		key[0] = byte(storage.STNEP11Transfers)
 	} else {
 		key[0] = byte(storage.STNEP17Transfers)
 	}
 	copy(key[1:], acc.BytesBE())
-	binary.LittleEndian.PutUint32(key[util.Uint160Size:], index)
+	binary.BigEndian.PutUint64(key[1+util.Uint160Size:], newestTimestamp)
+	binary.BigEndian.PutUint32(key[1+util.Uint160Size+8:], index)
 	return key
 }
 
+// SeekNEP17TransferLog executes f for each NEP-17 transfer in log starting from
+// the transfer with the newest timestamp up to the oldest transfer. It continues
+// iteration until false is returned from f. The last non-nil error is returned.
+func (dao *Simple) SeekNEP17TransferLog(acc util.Uint160, newestTimestamp uint64, f func(*state.NEP17Transfer) (bool, error)) error {
+	key := getTokenTransferLogKey(acc, newestTimestamp, 0, false)
+	prefixLen := 1 + util.Uint160Size
+	var seekErr error
+	dao.Store.Seek(storage.SeekRange{
+		Prefix:    key[:prefixLen],
+		Start:     key[prefixLen : prefixLen+8],
+		Backwards: true,
+	}, func(k, v []byte) bool {
+		lg := &state.TokenTransferLog{Raw: v}
+		cont, err := lg.ForEachNEP17(f)
+		if err != nil {
+			seekErr = err
+		}
+		return cont
+	})
+	return seekErr
+}
+
+// SeekNEP11TransferLog executes f for each NEP-11 transfer in log starting from
+// the transfer with the newest timestamp up to the oldest transfer. It continues
+// iteration until false is returned from f. The last non-nil error is returned.
+func (dao *Simple) SeekNEP11TransferLog(acc util.Uint160, newestTimestamp uint64, f func(*state.NEP11Transfer) (bool, error)) error {
+	key := getTokenTransferLogKey(acc, newestTimestamp, 0, true)
+	prefixLen := 1 + util.Uint160Size
+	var seekErr error
+	dao.Store.Seek(storage.SeekRange{
+		Prefix:    key[:prefixLen],
+		Start:     key[prefixLen : prefixLen+8],
+		Backwards: true,
+	}, func(k, v []byte) bool {
+		lg := &state.TokenTransferLog{Raw: v}
+		cont, err := lg.ForEachNEP11(f)
+		if err != nil {
+			seekErr = err
+		}
+		return cont
+	})
+	return seekErr
+}
+
 // GetTokenTransferLog retrieves transfer log from the cache.
-func (dao *Simple) GetTokenTransferLog(acc util.Uint160, index uint32, isNEP11 bool) (*state.TokenTransferLog, error) {
-	key := getTokenTransferLogKey(acc, index, isNEP11)
+func (dao *Simple) GetTokenTransferLog(acc util.Uint160, newestTimestamp uint64, index uint32, isNEP11 bool) (*state.TokenTransferLog, error) {
+	key := getTokenTransferLogKey(acc, newestTimestamp, index, isNEP11)
 	value, err := dao.Store.Get(key)
 	if err != nil {
 		if err == storage.ErrKeyNotFound {
@@ -206,8 +251,8 @@ func (dao *Simple) GetTokenTransferLog(acc util.Uint160, index uint32, isNEP11 b
 }
 
 // PutTokenTransferLog saves given transfer log in the cache.
-func (dao *Simple) PutTokenTransferLog(acc util.Uint160, index uint32, isNEP11 bool, lg *state.TokenTransferLog) error {
-	key := getTokenTransferLogKey(acc, index, isNEP11)
+func (dao *Simple) PutTokenTransferLog(acc util.Uint160, start uint64, index uint32, isNEP11 bool, lg *state.TokenTransferLog) error {
+	key := getTokenTransferLogKey(acc, start, index, isNEP11)
 	return dao.Store.Put(key, lg.Raw)
 }