Close #2322, close #2263.

TODO:

- [x] Apply and evaluate https://github.com/nspcc-dev/neo-go/commit/05b5fdc0f4e181777308b87f49433f3cab10be5d to improve GasPerVote retrieving a bit more.
- [x] Try to restore mainnet chain from dump and compare the resulting time.