@AnnaShaleva 2022-01-19
@roman-khimov, turns out that 05b5fdc makes things worse even if an early stop is available, I've checked this on reward distance up to 1000 blocks. Two `Seek` operations in a raw are more resource-demanding than one Seek with additional filtering logic for `start` and `end` heights. Although we can see that the longer reward distance is, the more effective 05b5fdc performs, so there's a win-tendency with larger reward distance.

All in all, I still doubt about 05b5fdc. Do we have a lot of cases in real life when reward distance is much more than 1000? If so then I may take a few additional benchmarks with adjusted reward distance and number of reward records. Current benchmark results look like:
```
name                                                           old time/op    new time/op    delta
NEO_GetGASPerVote/BoltPS_10RewardRecords_1RewardDistance-8       22.6µs ± 4%    41.9µs ± 3%  +85.42%  (p=0.000 n=9+10)
NEO_GetGASPerVote/BoltPS_10RewardRecords_10RewardDistance-8      24.8µs ± 4%    46.4µs ± 3%  +86.95%  (p=0.000 n=8+9)
NEO_GetGASPerVote/BoltPS_10RewardRecords_100RewardDistance-8     38.9µs ± 1%    68.2µs ±11%  +75.29%  (p=0.000 n=8+9)
NEO_GetGASPerVote/BoltPS_10RewardRecords_1000RewardDistance-8    84.9µs ± 9%    96.5µs ±18%  +13.73%  (p=0.006 n=9+9)
NEO_GetGASPerVote/BoltPS_100RewardRecords_1RewardDistance-8      22.1µs ±23%    34.1µs ±18%  +54.41%  (p=0.002 n=10+4)

name                                                           old alloc/op   new alloc/op   delta
NEO_GetGASPerVote/BoltPS_10RewardRecords_1RewardDistance-8       4.64kB ± 2%    8.31kB ± 2%  +78.97%  (p=0.000 n=9+10)
NEO_GetGASPerVote/BoltPS_10RewardRecords_10RewardDistance-8      4.75kB ± 5%    8.78kB ± 2%  +84.96%  (p=0.000 n=10+9)
NEO_GetGASPerVote/BoltPS_10RewardRecords_100RewardDistance-8     7.53kB ± 1%   12.44kB ± 5%  +65.21%  (p=0.000 n=9+9)
NEO_GetGASPerVote/BoltPS_10RewardRecords_1000RewardDistance-8    33.2kB ± 7%    37.6kB ± 9%  +13.42%  (p=0.000 n=10+9)
NEO_GetGASPerVote/BoltPS_100RewardRecords_1RewardDistance-8      5.84kB ±13%    8.99kB ± 9%  +53.83%  (p=0.002 n=10+4)

name                                                           old allocs/op  new allocs/op  delta
NEO_GetGASPerVote/BoltPS_10RewardRecords_1RewardDistance-8         81.7 ± 1%     128.0 ± 0%  +56.67%  (p=0.000 n=10+7)
NEO_GetGASPerVote/BoltPS_10RewardRecords_10RewardDistance-8        86.0 ± 0%     139.8 ± 1%  +62.53%  (p=0.000 n=8+9)
NEO_GetGASPerVote/BoltPS_10RewardRecords_100RewardDistance-8        139 ± 0%       202 ± 1%  +45.16%  (p=0.000 n=9+9)
NEO_GetGASPerVote/BoltPS_10RewardRecords_1000RewardDistance-8       789 ± 1%       861 ± 2%   +9.15%  (p=0.000 n=10+9)
NEO_GetGASPerVote/BoltPS_100RewardRecords_1RewardDistance-8        94.2 ± 3%     136.8 ± 4%  +45.14%  (p=0.003 n=9+4)
```