@roman-khimov 2022-01-19
> Two Seek operations in a raw are more resource-demanding than one Seek with additional filtering logic for start and end heights

That may be the case, that's why we're benchmarking all of these things, to see whether our assumptions are correct or not.

> Do we have a lot of cases in real life when reward distance is much more than 1000?

Yes. A day of blockchain life with 15s blocks is already 5760 blocks, people don't claim their GAS 6 times per day.

Theoretically we can have both ways of getting the value and choose between them based on the distance, but that might make the implementation a bit more complex than we'd like it to be.