@roman-khimov 2022-01-19
I think we can concentrate on things that reliably perform better now (the set we have at the moment looks like this), then spend a bit more time on additional tests/optimizations.