@AnnaShaleva 2022-01-19
> I think we can concentrate on things that reliably perform better now (the set we have at the moment looks like this), then spend a bit more time on additional tests/optimizations.

OK, and here's the last batch of results for 1 day, 3 days and a week of reward distance, no improvements:
```
name                                                              old time/op    new time/op    delta
NEO_GetGASPerVote/BoltPS_100RewardRecords_5760RewardDistance-8       212µs ±28%     262µs ± 7%  +23.60%  (p=0.016 n=5+5)
NEO_GetGASPerVote/BoltPS_100RewardRecords_17280RewardDistance-8      641µs ±29%     648µs ±26%     ~     (p=1.000 n=5+5)
NEO_GetGASPerVote/BoltPS_100RewardRecords_40320RewardDistance-8     1.29ms ± 2%    1.63ms ±46%  +26.29%  (p=0.008 n=5+5)
NEO_GetGASPerVote/LevelPS_100RewardRecords_5760RewardDistance-8      467µs ±12%     549µs ±23%     ~     (p=0.095 n=5+5)
NEO_GetGASPerVote/LevelPS_100RewardRecords_17280RewardDistance-8    1.13ms ±12%    1.08ms ± 9%     ~     (p=0.222 n=5+5)
NEO_GetGASPerVote/LevelPS_100RewardRecords_40320RewardDistance-8    2.66ms ±20%    2.46ms ± 7%     ~     (p=0.151 n=5+5)

name                                                              old alloc/op   new alloc/op   delta
NEO_GetGASPerVote/BoltPS_100RewardRecords_5760RewardDistance-8       127kB ± 4%     136kB ± 2%   +7.00%  (p=0.008 n=5+5)
NEO_GetGASPerVote/BoltPS_100RewardRecords_17280RewardDistance-8      401kB ± 6%     408kB ± 3%     ~     (p=0.690 n=5+5)
NEO_GetGASPerVote/BoltPS_100RewardRecords_40320RewardDistance-8      949kB ± 7%     950kB ± 8%     ~     (p=0.841 n=5+5)
NEO_GetGASPerVote/LevelPS_100RewardRecords_5760RewardDistance-8      144kB ± 3%     156kB ± 4%   +8.29%  (p=0.008 n=5+5)
NEO_GetGASPerVote/LevelPS_100RewardRecords_17280RewardDistance-8     411kB ± 2%     423kB ± 4%     ~     (p=0.222 n=5+5)
NEO_GetGASPerVote/LevelPS_100RewardRecords_40320RewardDistance-8     970kB ± 1%     988kB ± 2%   +1.90%  (p=0.032 n=5+5)

name                                                              old allocs/op  new allocs/op  delta
NEO_GetGASPerVote/BoltPS_100RewardRecords_5760RewardDistance-8       3.99k ± 1%     4.04k ± 0%   +1.09%  (p=0.008 n=5+5)
NEO_GetGASPerVote/BoltPS_100RewardRecords_17280RewardDistance-8      12.1k ± 1%     12.1k ± 1%     ~     (p=1.000 n=5+5)
NEO_GetGASPerVote/BoltPS_100RewardRecords_40320RewardDistance-8      28.7k ± 3%     28.5k ± 2%     ~     (p=0.690 n=5+5)
NEO_GetGASPerVote/LevelPS_100RewardRecords_5760RewardDistance-8      4.04k ± 0%     4.15k ± 1%   +2.90%  (p=0.008 n=5+5)
NEO_GetGASPerVote/LevelPS_100RewardRecords_17280RewardDistance-8     12.1k ± 1%     12.2k ± 1%     ~     (p=0.222 n=5+5)
NEO_GetGASPerVote/LevelPS_100RewardRecords_40320RewardDistance-8     28.2k ± 1%     28.3k ± 1%     ~     (p=0.548 n=5+5)
```