@roman-khimov 2022-01-19
> OK, and here's the last batch of results for 1 day, 3 days and a week of reward distance, no improvements:

OK, good. Negative result is a result also, now we know that this approach doesn't really help and for short distances it's just harmful. If that's the case, then we're done with #2322 also, the good part of it (early iteration stop) goes in, the bad part (double Seek) stays out.