@AnnaShaleva 2022-01-19
And also for the record, restoring mainnet up to 606249 height with default configuration, t took ~9 min. 35 sec. on master and ~7 min. 48 sec. on the current branch which is ~18% improvement.